src/requirements.txt
Flask==2.3.2
boto3==1.28.0
botocore==1.31.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0


